+++ src/Video/MultipleSourceVideoInterface.php
@@ -1,0 +1,14 @@
+<?php
+
+/*
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\VideoBundle\Video;
+
+interface MultipleSourceVideoInterface extends VideoInterface
+{
+    public function getMultipleSrc(): array;
+}